requirements.txt
numpy==1.19.2
Pillow==8.0.0
pure-python-adb==0.3.0.dev0
